csflask/config.py
import os
from decouple import config

class Config:
    SECRET_KEY = config('SECRET_KEY')
    SQLALCHEMY_DATABASE_URI = config('SQLALCHEMY_DATABASE_URI')
    MAIL_SERVER = 'smtp.googlemail.com'
    MAIL_PORT = 587
    MAIL_USE_TLS = True
    MAIL_USERNAME = config('USERNAME')
    MAIL_PASSWORD = config('PASSWORD')
    GUNMAIL_PASSWORD = config("GUNMAIL_PASSWORD")