requirements.txt
astroid==2.4.2
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-blueprint==1.3.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mime==0.1.0
Pillow==9.1.0
pycparser==2.20
pylint==2.6.0
python-decouple==3.3
six==1.15.0
# smtp==0.1.0
SQLAlchemy==1.3.20
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
